requirements.txt
Django==1.8.11
django-cors-headers==1.1.0
djangorestframework==3.3.2
